move utils to integration subdirectory
