+++ tests/integration/test_extract_docx.py
@@ -6,7 +6,7 @@ import pytest
 from nv_ingest_api.util.message_brokers.simple_message_broker import SimpleClient
 from nv_ingest_client.client import Ingestor
 from nv_ingest_client.client import NvIngestClient
-from tests.utilities_for_test import canonicalize_markdown_table
+from tests.integration.utilities_for_test import canonicalize_markdown_table
 
 
 @pytest.mark.integration
